moved everything to cog that has no other dependency
